EthTranchePricing: Derived TokenTranchePricing to handle eth instead of tokens
This is practically TokenTranchePricing just modified to handle weis and ether
tranches instead of token tranches.

When the functionality is approved, it might be a good idea to inherit from
TokenTranchePricing or make a common TranchePricingStrategy to inherit by both.

Also added relevant tests based on test_token_tranche_pricing.py.